+++ h2o-py/dynamic_tests/testdir_algos/kmeans/pyunit_kmeans_gridsearch_over_all_params_large.py
@@ -42,7 +42,7 @@ class Test_kmeans_grid_search:
     max_grid_model = 100           # maximum number of grid models generated before adding max_runtime_secs
 
     curr_time = str(round(time.time()))     # store current timestamp, used as part of filenames.
-    seed = round(time.time())
+    seed = int(round(time.time()))
 
     # parameters denoting filenames of interested
     training1_filenames = "smalldata/gridsearch/kmeans_8_centers_3_coords.csv"
