+++ h2o-py/dynamic_tests/testdir_algos/pca/pyunit_pca_gridsearch_over_all_params_large.py
@@ -43,7 +43,7 @@ class Test_PCA_grid_search:
     max_grid_model = 100           # maximum number of grid models generated before adding max_runtime_secs
 
     curr_time = str(round(time.time()))     # store current timestamp, used as part of filenames.
-    seed = round(time.time())
+    seed = int(round(time.time()))
 
     # parameters denoting filenames of interested that store training/validation/test data sets in csv format
     training1_filenames = "smalldata/gridsearch/pca1000by25.csv"
